rename_and_fill_readme
rename scripts
fill README file with installation, run and usage information
